fix(compose): measure through upstream's Measurer2
The port measured through `Measurer`, which upstream deprecated with
"Replace with Measurer2 instead for proper Measure/Layout handling."
The port had also dropped that @Deprecated annotation, so nothing
signalled the lag.

Moves the class into Measurer.kt as `Measurer2`, matching upstream's
file layout, and applies upstream's delta:

- `frameCache` is keyed by `anyOrNullId` rather than by the `Measurable`
  instance. This is what the change is for: under Lookahead the
  measurable is a different instance, and the old code worked around
  that by scanning the key set and re-measuring inside the placement
  scope.
- `placeables` is supplied per pass by the MeasurePolicy instead of
  being owned by the measurer, so the Lookahead and non-Lookahead
  passes no longer share one container.
- A missing frame now skips that one child (`return@fastForEach`)
  instead of abandoning placement of every remaining child (`return`).

Design elements and debug bounds are kept, deliberately. Upstream lost
them in this migration — there they survive only on the deprecated
`Measurer`, which nothing calls. Here `ConstraintLayout`'s JSON overload
still renders both, so following upstream exactly would have silently
removed a shipped feature.

All seven targets compile; 495 tests pass, including the parity suite.
