+++ compose/src/commonMain/kotlin/androidx/constraintlayout/compose/LateMotionLayout.kt
@@ -25,8 +25,10 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.Measurable
 import androidx.compose.ui.layout.MeasurePolicy
 import androidx.compose.ui.layout.MultiMeasureLayout
+import androidx.compose.ui.layout.Placeable
 import androidx.compose.ui.node.Ref
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.semantics.semantics
@@ -121,6 +123,7 @@ private fun lateMotionLayoutMeasurePolicy(
     // notifying our Composable caller
     contentTracker.value
 
+    val placeableMap = mutableMapOf<Measurable, Placeable>()
     val layoutSize =
         measurer.performInterpolationMeasure(
             constraints = constraints,
@@ -129,6 +132,7 @@ private fun lateMotionLayoutMeasurePolicy(
             constraintSetEnd = endProvider(),
             transition = TransitionImpl.EMPTY,
             measurables = measurables,
+            placeableMap = placeableMap,
             optimizationLevel = optimizationLevel,
             progress = motionProgress.value,
             compositionSource = compositionSource.value ?: CompositionSource.Unknown,
@@ -136,5 +140,7 @@ private fun lateMotionLayoutMeasurePolicy(
         )
     compositionSource.value = CompositionSource.Unknown // Reset after measuring
 
-    layout(layoutSize.width, layoutSize.height) { with(measurer) { performLayout(measurables) } }
+    layout(layoutSize.width, layoutSize.height) {
+        with(measurer) { performLayout(measurables = measurables, placeableMap = placeableMap) }
+    }
 }
