+++ compose/src/commonMain/kotlin/androidx/constraintlayout/compose/MotionLayout.kt
@@ -39,8 +39,10 @@ import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.geometry.Rect
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.Measurable
 import androidx.compose.ui.layout.MeasurePolicy
 import androidx.compose.ui.layout.MultiMeasureLayout
+import androidx.compose.ui.layout.Placeable
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.onPlaced
 import androidx.compose.ui.node.Ref
@@ -981,6 +983,7 @@ internal fun motionLayoutMeasurePolicy(
     // notifying our Composable caller
     contentTracker.value
 
+    val placeableMap = mutableMapOf<Measurable, Placeable>()
     val layoutSize =
         measurer.performInterpolationMeasure(
             constraints = constraints,
@@ -989,6 +992,7 @@ internal fun motionLayoutMeasurePolicy(
             constraintSetEnd = constraintSetEnd,
             transition = transition,
             measurables = measurables,
+            placeableMap = placeableMap,
             optimizationLevel = optimizationLevel,
             progress = motionProgress.floatValue,
             compositionSource = compositionSource.value ?: CompositionSource.Unknown,
@@ -996,7 +1000,9 @@ internal fun motionLayoutMeasurePolicy(
         )
     compositionSource.value = CompositionSource.Unknown // Reset after measuring
 
-    layout(layoutSize.width, layoutSize.height) { with(measurer) { performLayout(measurables) } }
+    layout(layoutSize.width, layoutSize.height) {
+        with(measurer) { performLayout(measurables = measurables, placeableMap = placeableMap) }
+    }
 }
 
 /**
