+++ compose/src/commonMain/kotlin/androidx/constraintlayout/compose/MotionMeasurer.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.graphics.PathEffect
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import androidx.compose.ui.graphics.drawscope.translate
 import androidx.compose.ui.layout.Measurable
+import androidx.compose.ui.layout.Placeable
 import androidx.compose.ui.layout.layoutId
 import androidx.compose.ui.unit.Constraints
 import androidx.compose.ui.unit.Density
@@ -38,7 +39,7 @@ import androidx.constraintlayout.core.state.WidgetFrame
 import androidx.constraintlayout.core.widgets.Optimizer
 
 @ExperimentalMotionApi
-internal class MotionMeasurer(density: Density) : Measurer(density) {
+internal class MotionMeasurer(density: Density) : Measurer2(density) {
     private val DEBUG = false
     private var lastProgressInInterpolation = 0f
     val transition = Transition { with(density) { it.dp.toPx() } }
@@ -81,11 +82,13 @@ internal class MotionMeasurer(density: Density) : Measurer(density) {
         constraintSetEnd: ConstraintSet,
         @SuppressWarnings("HiddenTypeParameter") transition: TransitionImpl,
         measurables: List<Measurable>,
+        placeableMap: MutableMap<Measurable, Placeable>, // Initialized by caller, filled by us
         optimizationLevel: Int,
         progress: Float,
         compositionSource: CompositionSource,
         invalidateOnConstraintsCallback: ShouldInvalidateCallback?,
     ): IntSize {
+        placeables = placeableMap
         val needsRemeasure =
             needsRemeasure(
                 constraints = constraints,
@@ -225,7 +228,7 @@ internal class MotionMeasurer(density: Density) : Measurer(density) {
                 measurable.measure(
                     Constraints.fixed(interpolatedFrame.width(), interpolatedFrame.height()),
                 )
-            frameCache[measurable] = interpolatedFrame
+            frameCache[measurable.anyOrNullId] = interpolatedFrame
         }
 
         if (layoutInformationReceiver?.getLayoutInformationMode() == LayoutInfoFlags.BOUNDS) {
